Re-develop BackdropBottomSheetBehavior
Re-develop BackdropBottomSheetBehavior to avoid using previous Y value and multiplier and use real time values instead. Make the behavior unaware of other behaviors. Add attributes for peekValue. We still need to work on the case someone want to change the peekHeight programmaticaly.
This way it fix strange behavior where the ImageView attached to this behavior can be detached from the (GoogleStyle)BottomSheetBehavior depending on the strength of the sliding gesture. In addition it become reusable without any (GoogleStyle)BottomSheetBehavior !