Make BackDropBottomSheet work after screen rotation
Rely on the dependency height in order to calculate the collapsedY instead of the Y position of the dependency. In case of the screen rotation when the bottomsheet is expanded the Y position of the dependency cannot permit to calculate the collapsedY value. The collapsedY is equal to the size of the dependency less (the Peak Height plus status bar plus toolbar -> roughly 2 time the Peak Height)